requirements.txt
ultralytics==8.0.178
pandas==2.2.2
flask==3.0.2
pyyaml==6.0.1
